# backend/queues/api-docs.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Queues Service API",
    "description" : "Contains operations for working with Items, Queues and Users",
    "version" : "SNAPSHOT"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "default relative url"
  } ],
...
  "paths" : {
    "/api/dictionary/{type}" : {
      "get" : {
        "tags" : [ "dictionary" ],
        "summary" : "Get relevant results for specific type",
        "operationId" : "getRelevantDictionaryEntities",
        "parameters" : [ {
          "name" : "type",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "value",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "oneOf" : [ {
                    "$ref" : "#/components/schemas/ErrorDTO"
                  }, {
                    "type" : "object"
                  } ]
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "oneOf" : [ {
                    "$ref" : "#/components/schemas/ErrorDTO"
                  }, {
                    "type" : "object"
                  } ]
                }
              }
            }
          },
          "422" : {
            "description" : "Unprocessable Entity",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorDTO"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorDTO"
                }
              }
            }
          },
          "429" : {
            "description" : "Too Many Requests",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorDTO"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "oneOf" : [ {
                    "$ref" : "#/components/schemas/ErrorDTO"
                  }, {
                    "type" : "object"
                  } ]
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "oneOf" : [ {
                    "$ref" : "#/components/schemas/ErrorDTO"
                  }, {
                    "type" : "object"
                  } ]
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "mr_user_auth" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "dictionary" ],
        "summary" : "Creates a dictionary entry for specific type",
        "operationId" : "createDictionaryEntity",
        "parameters" : [ {
          "name" : "type",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DictionaryValueDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "oneOf" : [ {
                    "$ref" : "#/components/schemas/ErrorDTO"
                  }, {
                    "type" : "object"
                  } ]
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "oneOf" : [ {
                    "$ref" : "#/components/schemas/ErrorDTO"
                  }, {
                    "type" : "object"
                  } ]
                }
              }
            }
          },
          "422" : {
            "description" : "Unprocessable Entity",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorDTO"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorDTO"
                }
              }
            }
          },
          "429" : {
            "description" : "Too Many Requests",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorDTO"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "oneOf" : [ {
                    "$ref" : "#/components/schemas/ErrorDTO"
                  }, {
                    "type" : "object"
                  } ]
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "oneOf" : [ {
                    "$ref" : "#/components/schemas/ErrorDTO"
                  }, {
                    "type" : "object"
                  } ]
                }
              }
            }
          },
          "200" : {
            "description" : "OK"
          }
        },
        "security" : [ {
          "mr_user_auth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ErrorDTO" : {
        "required" : [ "description", "time" ],
        "type" : "object",
        "properties" : {
          "time" : {
            "type" : "string",
            "description" : "The time of error generation",
            "format" : "date-time"
          },
          "description" : {
            "type" : "string",
            "description" : "The common error description"
          },
          "details" : {
            "type" : "object",
            "description" : "The details for custom troubleshooting. Could have any format or be not presented",
            "nullable" : true
          }
        }
      },
      "DictionaryValueDTO" : {
        "required" : [ "value" ],
        "type" : "object",
        "properties" : {
          "value" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "mr_user_auth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://login.microsoftonline.com/74143e14-36bb-4cd2-bdb9-65d0f8b5b360/oauth2/authorize?resource=94021731-2fa2-4ee1-a512-af52d3b7a552",
            "scopes" : { }
          }
        }
      }
    }
  }
}